extract datastructures from CompiledSouffle
